AuthoritiesTokenExtractor to standalone class
Extracted AuthoritiesTokenExtractor into its own class for better reusability and code maintainability. Updated related logic in CoreFunctionService to utilize the standalone implementation. It helps authorities to be accessed in expressions.
